requirements.txt
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.4
